:sparkles: Don't exit the shell when Ctrl-D is pressed
Because I tend to work in a multi-terminal-window setup, it's more common
that I think I'm Ctrl-Ding my way out of some REPL but I'm looking at the
wrong window, and so exit a shell (and close the window), than I Ctrl-D out
of a shell on purpose. I'd normally "exit RET" my way out of an actual
shell.

So, given this annoys me far more than it is useful.... let's nuke the
Ctrl-D binding as a method of getting out of a shell.