:racehorse: Have the prompt come back faster
Just updated xcode and, after doing so, I noticed that the prompt was a lot
slower coming back, a very obvious faction of a second to return after
executing a command.

A quick binary chop at the highest level of the prompt code showed that it
was something in the svn information functions. I don't use svn, never
have...